% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{add_in_area}
\alias{add_in_area}
\title{Take set count data for a species and add column indicating whether or not
it is inside a given area}
\usage{
add_in_area(set_counts_of_sp, area = HG_herring_pred_area)
}
\arguments{
\item{set_counts_of_sp}{input tibble of the species of interest, likely generated from
\code{cache_pbs_data_iphc} (which seems to save a list with the first element
the desired tibble), with column names \code{year}, \code{station}, \code{lat}, \code{lon},
\code{E_it}, \code{N_it}, \code{C_it}, \code{E_it20}, \code{N_it20}, \code{C_it20}, \code{usable}, and
\code{in_area} if \code{indicate_in_area} is TRUE}

\item{area}{\code{data.frame} of (PBSmapping) class \code{PolySet} with column names
\code{PID}, \code{SID}, \code{POS}, \code{X}, \code{Y},}
}
\value{
tibble of \code{set_counts_of_sp} with extra logical column \code{in_area} to
say whether or not each set is within \code{area}. Internally a column \code{EID} is
created to assign each set-year a unique ID, needed for
\code{PBSmapping::findPolys}, but this is not returned.
}
\description{
See example and the vignette.
}
\examples{
\dontrun{
sp_set_counts_with_area <- add_in_area(yelloweye_rockfish$set_counts,
                                       HG_herring_pred_area)
# Gives a warning about hole vertices, can ignore.
}
}
\author{
Andrew Edwards
}
