% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-maps.R
\name{add_stations}
\alias{add_stations}
\title{Add individual stations to map}
\usage{
add_stations(
  set_counts_of_sp_one_year,
  species = TRUE,
  pch_zero_count,
  pch_pos_count,
  cex_val
)
}
\arguments{
\item{set_counts_of_sp_one_year}{tibble for just one year, from
\code{set_counts_of_sp}.}

\item{species}{TRUE to plot the stations and catch rates for the given
species (already specified in the data), or FALSE to just low locations of
stations and whether they are usable or not.}

\item{usable}{TRUE if plotting the usable stations, FALSE for unusable.}

\item{pos_catch_rates}{TRUE if plotting the positive (>0) catch rates for
the given species, FALSE if plotting the 0 catch rates (showing as open
circles).}
}
\value{
adds points to existing plot
}
\description{
Add individual stations to map of BC, with options for all stations (no
species catch information), or positive and zero catches for a given
species. Called from within \code{plot_iphc_map()}.
}
\author{
Andrew Edwards
}
