% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{boot_iphc}
\alias{boot_iphc}
\title{Do bootstrap calculations on each IPHC Series}
\usage{
boot_iphc(ser_year_rates, num.boots = 1000, seed_val = 42)
}
\arguments{
\item{ser_year_rates}{Tibble of two columns containing just year and catch
rates (either C_it or C_it20 depending on the Series).}

\item{num.boots}{Number of bootstrap replicates to do.}

\item{seed_val}{Seed for random number generator.}
}
\value{
Tibble containg one row for each year, with columns year, I_tBootMean,
I_tBootLow, I_tBootHigh and I_tBootCV. If ser_year_rates is NULL or all
counts are NA then return year = 2003 and all NA's (to not break later code).
}
\description{
To be called from within \code{\link[=calc_iphc_ser_all]{calc_iphc_ser_all()}},
once for each series.
}
