% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{calc_iphc_ser_all}
\alias{calc_iphc_ser_all}
\title{Calculate Series A, B, C and D from the IPHC data}
\usage{
calc_iphc_ser_all(set_counts, lat_cut_off = 50.6)
}
\arguments{
\item{set_counts}{species-specific set-level counts from \code{\link[=tidy_iphc_survey]{tidy_iphc_survey()}}
or other.}

\item{lat_cut_off}{cut off below which sets are excluded for Series A and B,
default is that used in YYR 2014 assessment.}
}
\value{
list containing four tibbles, one for each survey (ser_A, ser_B, ser_C
and ser_D). Each tibble has one row for each set in each year, with columns
year, station, lat, lon, E_it (effective skate number), N_it (number of
fish caught of the given species), C_it (catch rate of given species, as
numbers per effective skate), E_it20, N_it20 and C_it20 are the same but
considering the first 20 hooks of each skate only, usable (whether the set
should be used, based on IPHC codes; note that some should not be used for
geospatial analysis but are included here).
}
\description{
Calculate all four series for as many years as possible, including bootstrapped
values. Series with no data are treated slightly different to each other
(ser_C ends up as an empty tibble, which may cause problems down the road).
}
\details{
The four series are:

Series A: first 20 hooks from each skate, only north of WCVI

Series B: all hooks from each skate, only north of WCVI

Series C: all hooks from each skate, full coast

Series D: first 20 hooks from each skate, full coast
}
\examples{
\dontrun{
yelloweye <- tidy_iphc_survey(
 get_iphc_hooks("yelloweye rockfish"),
 get_iphc_skates_info(),
 get_iphc_sets_info()
)
calc_iphc_ser_all(yelloweye)
}
}
