% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{calc_iphc_ser_E_and_F}
\alias{calc_iphc_ser_E_and_F}
\title{Calculations restricted to a user-defined area.}
\usage{
calc_iphc_ser_E_and_F(set_counts)
}
\arguments{
\item{set_counts}{species-specific set-level counts from \code{\link[=tidy_iphc_survey]{tidy_iphc_survey()}}
or other, but with extra column \code{in_area} indicating if each station is in
the area being considered or not. See vignette for example.}
}
\value{
list of class \code{IPHC_ser_E_and_F} containing two tibbles, one for each series (ser_E and ser_F).
Each tibble has one row for each set in each year, restricted to only the
sets within the defined area, with columns
year, station, lat, lon, E_it (effective skate number), N_it (number of
fish caught of the given species), C_it (catch rate of given species, as
numbers per effective skate), E_it20, N_it20 and C_it20 are the same but
considering the first 20 hooks of each skate only, usable (whether the set
should be used, based on IPHC codes; note that some should not be used for
geospatial analysis but are included here).
}
\description{
Calculate Series E and F from the IPHC data (restricted to a given area)
}
\details{
Calculate both Series for as many years as possible, including bootstrapped
values, for the data restricted to a user-defined area.

The two series are:

Series E: first 20 hooks from each skate (like Series A but restricted to a
given area)

Series F: all hooks from each skate (like Series B but restricted to a given area)
}
\examples{
\dontrun{
TODO:
yelloweye <- tidy_iphc_survey(
 get_iphc_hooks("yelloweye rockfish"),
 get_iphc_skates_info(),
 get_iphc_sets_info()
)
calc_iphc_ser_all(yelloweye)
}
}
