% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{calc_iphc_ser_EF}
\alias{calc_iphc_ser_EF}
\title{Calculate Series EF and test if the scaled E and F are significantly different}
\usage{
calc_iphc_ser_EF(series_all, return_F_if_same_length = TRUE)
}
\arguments{
\item{series_all}{List of tibbles, one for each of Series E and F,
resulting from \code{calc_iphc_ser_E_and_F()}}

\item{return_F_if_same_length}{logical to return Series F if Series E and F
have the same number of years. User should check which is more suitable and
then change this value if necessary (F uses all hooks but E includes 1995
and 1996)}
}
\value{
List containing

ser_longest: the longest time series possible from
Series E and F,

test_EF: the results from the paired t-test, NULL if Series F is longest.

G_E, G_F: geometric means of nonzero values in Series E and Series F
(based on bootstrapped means).

Longest series is either

(i) Series EF (Series E with 1995 and 1996 appropriately scaled from
Series F) if \code{test_EF$p.value >= 0.05}, because the p-value means that we
cannot reject the null hypothesis that the true difference
in means of the rescaled (by their geometric means) Series E and F equals 0,

(ii) Series E, which is the longest series available if the rescaled
series are significantly different.

(iii) Series F, if the years for which only the first 20 hooks are enumerated
never caught the species (but other years do). Then Series EF becomes Series
E plus some zeros, so we may as well use Series F that uses all the hooks
and so is more likely to catch the species. This is a rare situation (and
likely arises because not all species were specifically identified in
earlier years (or 2013?), but seems to occur for China Rockfish for
Series AB analysis. Return NULL
for t_EF (to then use later to identify that Series F is the longest
series).

(iv) Series F (since all hooks) if E and F contain the same number of
years, but user can change \code{return_F_if_same_length} as Series E may make
more sense (since goes back to 1995). If E and F contain the same exact
years then F is always returned.

See vignette for example.
}
\description{
Calculate Series EF (Series E with 1995 and 1996 appropriately scaled from
Series F)
}
