% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{compare_iphc_ser_A_D}
\alias{compare_iphc_ser_A_D}
\title{Compare Series A and D to see if A can be considered a relative index for the whole coast}
\usage{
compare_iphc_ser_A_D(series_all)
}
\arguments{
\item{series_all}{List of tibbles, one for each of Series A, B, C and D,
resulting from \code{\link[=calc_iphc_ser_all]{calc_iphc_ser_all()}}}
}
\value{
List of t_AD (results of the paired t-test), and geometric means G_A
and G_D. t_AD is NULL if there is no catch in the overlapping years for
Series A or Series D.
}
\description{
Compare Series A and D to see if A can be considered a relative index for
the whole coast.
}
