% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countData2020}
\alias{countData2020}
\title{Count data for 20202 IPHC survey}
\format{
A tibble:
\describe{

\item{year}{2020 for this data set}
\item{station}{station name}
\item{spNameIPHC}{species name for the counts for that set}
\item{specCount}{count of that species for that set}
}
}
\source{
See \code{data-raw/iphc-2020-data.Rmd} for full details and how data were
extracted from IPHC website.
}
\usage{
countData2020
}
\description{
A dataset containing counts of each species for each
station for the 2020 IPHC survey.
}
\keyword{datasets}
