% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countData2021}
\alias{countData2021}
\title{Count data for 2021 IPHC survey (following 2020 approach since 20-hooks again)}
\format{
A tibble:
\describe{

\item{year}{2021 for this data set}
\item{station}{station name}
\item{spNameIPHC}{species name for the counts for that set}
\item{specCount}{count of that species for that set}
}
}
\source{
See \code{data-raw/iphc-2021-data.Rmd} and saved
\code{data-raw/iphc-2021-data.pdf} for full details and how data were
extracted from IPHC website.
}
\usage{
countData2021
}
\description{
A dataset containing counts of each species for each
station for the 2021 IPHC survey.
}
\keyword{datasets}
