% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{format_iphc_longest}
\alias{format_iphc_longest}
\title{Format the longest IPHC time series index to agree with other surveys}
\usage{
format_iphc_longest(iphc_set_counts_sp)
}
\arguments{
\item{iphc_set_counts_sp}{Output from \code{\link[=calc_iphc_full_res]{calc_iphc_full_res()}} (only actually
need the ser_longest component and test_AB).}
}
\value{
Renamed ser_longest, with some required columns calculated.
}
\description{
Format the longest IPHC time series index to agree with other surveys so
that plot_survey_index() in gfplot (and may get moved to here) works automatically. So the mean catch rate
gets renames as `biomass' even though it's numbers per effective skate.
}
\examples{
\dontrun{
# If already loaded data via gfsynopsis then
dat_iphc <- readRDS("../gfsynopsis/report/data-cache/iphc/yelloweye-rockfish.rds")
set_counts <- dat_iphc$set_counts
# Else to load from scratch:
# set_counts <- get_all_iphc_set_counts("yelloweye rockfish")
iphc_set_counts_sp <- calc_iphc_full_res(set_counts)
format_iphc_longest(iphc_set_counts_sp)
# Has no data for early years or 2013:
sp = "china rockfish"
# If already loaded data via gfsynopsis then
dat_iphc <- readRDS("../gfsynopsis/report/data-cache/iphc/china-rockfish.rds")
set_counts <- dat_iphc$set_counts
# Else to load from scratch:
# set_counts <- get_all_iphc_set_counts(sp)
iphc_set_counts_sp <- calc_iphc_full_res(set_counts)
format_iphc_longest(iphc_set_counts_sp)
}
}
