% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-data.R
\name{get_all_iphc_set_counts}
\alias{get_all_iphc_set_counts}
\title{Get and combine all the IPHC survey data for a given species}
\usage{
get_all_iphc_set_counts(species)
}
\arguments{
\item{species}{Species common name (as used by gfplot), or 'hook with bait'.}
}
\value{
Tibble contains year, station name, lat, lon,
E_it (effective skate number for that station, based on all
hooks),
N_it (number of 'species' caught on all hooks)
C_it (catch rate of 'species' as number per effective skate,
based on all hooks)
E_it20 (effective skate number for that station, based on first
20 hooks),
N_it20 (number of 'species' caught in first 20 hooks),
C_it20 (catch rate of 'species' as number per effective skate,
based on the first 20 hooks),
usable (whether or not that station is usable, as deemed by IPHC),
standard (whether or not station is a standard one or in the
expansion set after 2018).
}
\description{
Combine all the IPHC survey data for a given species, with catch rates
calculated
based on effective skates and numbers of hooks observed. Catch rates are
given for the first 20 hooks and/or for all hooks, depending on the data
collected.
}
