% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-data.R
\name{get_combined_species}
\alias{get_combined_species}
\title{Get combined catch counts for a set of species}
\usage{
get_combined_species(
  sp_vec,
  save_RDS_name = NULL,
  path = ".",
  compress = FALSE
)
}
\arguments{
\item{sp_vec}{vector of IPHC species names}

\item{save_RDS_name}{name (with .rds suffix) to save the .rds file of the
output (there is no natural obvious default based on \code{sp_vec}), if NULL
then do not save}

\item{path}{The folder where the cached data will be saved.}

\item{compress}{Compress the \code{.rds} file? Defaults to \code{FALSE} for faster
reading and writing at the expense of disk space.}
}
\value{
list with  tibble \code{set_counts} (same format as for individual
species) that contains year, station name, lat, lon, and
\itemize{
\item E_it - effective skate number for that station, based on all hooks (so
NA for 1997-2002, 2013, 2020),
\item N_it...<num> - number of 'species' caught on all hooks, so NA for
1997-2002, 2013, 2020, where <num> is the column number appended by dplyr within this
function,
\item N_it_sum - sum of \verb{N_it...<num>}, with NA's removed (treated as zeros)
TODO: say what happens if some have NA
\item C_it_sum - catch rate of all 'species' as number per effective skate,
based on all hooks, so NA for 1997-2002, 2013, 2020,
\item E_it20 - effective skate number for that station, based on first 20
hooks, so NA for 1995 and 1996,
\item N_it20...<num number of 'species' caught in first 20 hooks, so NA for
1995 and 1996, where <num> is the column number appended by dplyr within this
\item N_it20_sum - sum of \verb{N_it20...<num>}, with NA's removed (treated as zeros)
TODO: say what happens if some have NA
\item C_it20_sum - catch rate all 'species' as number per effective skate,
based on the first 20 hooks, so NA for 1995 and 1996,
\item usable  - whether or not that station is usable, as deemed by IPHC,
\item standard - whether or not station is a standard one or in the
expansion set after 2018; all Y here since pre-2018; for
1996 and 1997 the station layout is not the standard grid
and the naming of stations is different,
\verb{intersect(filter(sp_set_counts$set_counts, year == 1996)$station, setDataExpansion$station}
is empty, but the area fished is similar to later standard
years so we put Y here
}

If no data on any of the species in \code{sp_vec} then \code{C_it_all} and N_it are
NA's. This happens for sure for years when the counts are not
available. Examine each individual species count for NA's to check they
have propagated through okay.
}
\description{
The .rds file for each species needs to already exist, from running
\code{cache_pbs_data_iphc(sp_vec)}.
This combines the values together, for example to give data for a number of
skates combined, rather than individual species.
}
\author{
Andrew Edwards
}
