% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiphc.R
\docType{package}
\name{gfiphc-package}
\alias{gfiphc}
\alias{gfiphc-package}
\title{gfiphc: Data Extraction and Analysis for Groundfish Data from the IPHC Longline Survey in BC}
\description{
Facilitates International Pacific Halibut Commission (IPHC) survey groundfish data extraction at the Pacific Biological Station (PBS), Department of Fisheries and Oceans Canada, and analysis of the data for all species caught in British Columbia waters.
}
\author{
\strong{Maintainer}: Andrew M. Edwards \email{a@a.com}

Authors:
\itemize{
  \item Sean C. Anderson
  \item Elise A. Keppel
  \item Chris J. Grandin
}

}
\keyword{internal}
