% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HG_herring_pred_area}
\alias{HG_herring_pred_area}
\title{Area considered for predators of Haida Gwaii Pacific Herring}
\format{
A \code{data.frame}, also of class \code{PolySet} (from PBSmapping) with
columns, of which X and Y (longitude and latitude) are then used to
determine which IPHC stations are inside the relevant area:
\describe{
\item{PID}{primary identifier}
\item{SID}{secondary identify}
\item{POS}{not sure what this is}
\item{X}{longitude of station}
\item{Y}{latitude of station}
}
}
\source{
Jennifer Boldt sent Andrew Edwards the shape files for the area
(corresponding to those used for Pacific Hake predators of Pacific
Herring), which Rowan Haigh helped use with \code{PBSmapping} (primarily helping
overdrawing of islands) There are still some minor issues and extra lines
shown, possibly to do with islands on lakes on islands in the sea. Details
in \code{data-raw/herring-predators-HG/herring-predators-HG-area.R}.
}
\usage{
HG_herring_pred_area
}
\description{
Primarily to determine which stations are within the relevant area to then
generate indices of abundance for groundfish predators of Haida Gwaii Pacific
Herring. See vignette.
}
\keyword{datasets}
