% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hooks_with_bait}
\alias{hooks_with_bait}
\title{Available counts of hooks returned with bait on them, for each set for all years}
\format{
A \code{list} containing a single \code{tibble} called \code{set_counts}, with
columns as described in \code{?get_all_iphc_set_counts}.
}
\source{
Extracted from GFBio using \code{data-raw/sets-skates-hooks-yelloweye.R}.
}
\usage{
hooks_with_bait
}
\description{
Extracts and calculates all avaiable counts, at the set level, of hooks
returned with bait (which is essentially equivalent to a species).
}
\keyword{datasets}
