% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{iphc-spp-names.csv}
\alias{iphc-spp-names.csv}
\title{Conversion from GFbio names to IPHC species names (for data not in GFbio)}
\format{
A csv file:
\describe{
\item{species_common_nameyear}{GFbio species common name}
\item{iphc_common_name}{IPHC data sets species common name}
}
}
\source{
Created manually by Andrew Edwards. Most are obvious. If get a new
set-level dataset from IPHC (that doesn't get put into GFbio) then see
\code{\link[=check_iphc_spp_name]{check_iphc_spp_name()}}.
Currently this only has Type A species \strong{\strong{Andy to update with Type B.
\strong{And may also want "Hook with Bait" if different in GFbio}}}
}
\description{
A data fraom with the GFbio name and the corresponding name used by IPHC in
the datasets that are at the set level and so not extracted from GFbio.
}
