% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{iphc_format_for_EAFM}
\alias{iphc_format_for_EAFM}
\title{Format results from multiple species for Jennifer's EAFM HG Herring Case
Study and save them.}
\usage{
iphc_format_for_EAFM(
  sp_vec,
  sp_vec_list,
  path = ".",
  filename = "herring-HG-predators-IPHC.csv"
)
}
\arguments{
\item{sp_vec}{A vector of species names. Each one \emph{must} have a corresponding
list \code{species_name_area} in the workspace which is an output from
\code{iphc_get_calc_plot_area()} -- see vignette.}

\item{sp_vec_list}{list of output, with each element corresponding to each
species (and named for that species), and being a tibble of the longest
time series. See vignette, since needs to be created outside of a function.}

\item{path}{The folder where the cached results will be saved.}

\item{filename}{filename (including .csv) to save .csv file; also saves .rds
of sp_vec_list.}
}
\value{
saves a .csv file and returns the resulting tibble. Columns are:
\itemize{
\item Year
\item Indicator (e.g. ArrowtoothFlounder_IPHC_SurveyCatchRate,
ArrowtoothFlounder_IPHC_UpperCI)
\item Value
}
}
\description{
Create and save a .csv file in the format requested by Jennifer Bolt:
What would work best for me is a csv file with 3 columns: \code{Year},
\code{Indicator}, \code{Value}
Non-standardized values are best (i.e. not normalised).
The \code{Indicator} column could include:
\enumerate{
\item ArrowtoothFlounder_IPHC_SurveyCatchRate
\item ArrowtoothFlounder_IPHC_UpperCI
\item ArrowtoothFlounder_IPHC_LowerCI
\item Rockfish_IPHC_SurveyCatchRate
\item ...
}
}
\details{
Also saves \code{sp_vec_list} as an .rds file.
}
\author{
Andrew Edwards
}
