% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{iphc_get_calc_plot_area}
\alias{iphc_get_calc_plot_area}
\title{Get data, do calculations and plot longest series for the IPHC survey
restricted to a specified area}
\usage{
iphc_get_calc_plot_area(
  sp,
  area = HG_herring_pred_area,
  cached_data = TRUE,
  cached_results = FALSE,
  verbose = FALSE,
  print_sp_name = TRUE,
  path_data = ".",
  path_results = NULL
)
}
\arguments{
\item{sp}{Species name (as used in gfdata and gfplot). Or something like
"skates combined" -- see vignette.}

\item{area}{\code{data.frame} of (PBSmapping) class \code{PolySet} with column names
\code{PID}, \code{SID}, \code{POS}, \code{X}, \code{Y},}

\item{cached_data}{if TRUE then use cached data (path_data/sp-name.rds)}

\item{cached_results}{if TRUE then use cached results
(path_results/sp-name-results.rds), else do calculations here and save results}

\item{verbose}{if TRUE then print out some of the data (useful in vignette loops)}

\item{print_sp_name}{if TRUE then print out species name (useful in vignette
loops)}

\item{path_data}{path to save or load the cached data}

\item{path_results}{path to save or load the cached data}
}
\value{
For the given species, list containing

sp_set_counts_with_area: tibble returned from \code{\link[=add_in_area]{add_in_area()}} of set
counts of the species with with extra logical column \code{in_area} to
say whether or not each set is within \code{area}.

ser_E_and_F: list of class \code{IPHC_ser_E_and_F} containing two tibbles, one
for each series (ser_E and ser_F), as output from \code{\link[=calc_iphc_ser_E_and_F]{calc_iphc_ser_E_and_F()}}.

series_longest: list containing \code{ser_longest}, the longest possible time
series from Series E and F, and other objects, as output form \code{\link[=calc_iphc_ser_EF]{calc_iphc_ser_EF()}}.
}
\description{
Get data, do calculations and plot longest series for the IPHC survey for
a given species restricted to a given area. Will take a while if queries
GFbio (for which need to be on DFO network), else can use cached data
(species-name.rds is always the full data for the species, before any
calculations).
If data are for combined species, as cached with \code{get_combined_species()},
then the column \code{N_it_sum} etc. column names are used for \code{N_it} etc.
}
\examples{
\donttest{
iphc_get_calc_plot_area("yelloweye rockfish")
##' # Gives a warning about hole vertices, can ignore.
}
}
\author{
Andrew Edwards
}
