% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-maps.R
\name{plot_BC}
\alias{plot_BC}
\title{Map for BC coast, adapted from an example in Rowan Haigh's TESA course}
\usage{
plot_BC(
  xlim = c(-134, -124),
  ylim = c(48, 54.6),
  zlev = seq(200, 1200, 200),
  showBathymetry = FALSE,
  ...
)
}
\arguments{
\item{xlim}{range of x (longitude)}

\item{ylim}{range of y (latitude)}

\item{zlev}{depth contours to show}

\item{showBathymetry}{if TRUE then show bathymetry contours}

\item{...}{other arguments to \code{PBSmapping::plotMap()} or \code{PBSmapping::addLines()}}
}
\value{
Map of BC land and sea
}
\description{
Map for BC coast, adapted from an example in Rowan Haigh's TESA course
}
\author{
Andrew Edwards
}
