% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-maps.R
\name{plot_iphc_map_panel}
\alias{plot_iphc_map_panel}
\title{Plot a panel of maps}
\usage{
plot_iphc_map_panel(
  set_counts_of_sp,
  sp_short_name,
  years_to_show,
  par_mfrow = c(2, 2),
  ...
)
}
\arguments{
\item{set_counts_of_sp}{see \code{plot_iphc_map()}}

\item{sp_short_name}{see \code{plot_iphc_map()}}

\item{years_to_show}{see vector of years to show one map for each year}

\item{par_mfrow}{two-value vector (rows by columns) for \code{par(mfrow = par_mfrow)}}

\item{...}{further arguments to \code{plot_iphc_map()}}
}
\value{

}
\description{
Plot a panel of maps
}
\author{
Andrew Edwards
}
