% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-iphc.R
\name{plot_IPHC_ser_ABCD}
\alias{plot_IPHC_ser_ABCD}
\title{Single plotting function for Series A, B, C, D, or combination of two}
\usage{
plot_IPHC_ser_ABCD(
  series_ABCD_full,
  plot_type = "AB",
  x_lim = c(1995, as.integer(format(Sys.Date(), "\%Y")) - 1),
  y_lim = NULL,
  shift = 0.15,
  tck_length = -0.02,
  ser_A_col = "blue",
  ser_B_col = "red",
  ser_C_col = "orange",
  ser_D_col = "darkgreen",
  legend_text = NULL,
  x_lab = "Year",
  y_lab = "Catch rate index (numbers per effective skate)",
  gap_ci = 0.5,
  ...
)
}
\arguments{
\item{series_ABCD_full}{list containing tibbles for \code{ser_longest} etc.,
an output from \code{calc_iphc_full_res}}

\item{plot_type}{one of
\itemize{
\item \code{X}, where \code{X} is \code{A}, \code{B}, \code{C}, or \code{D}, to plot just Series X
\item \code{X_Y_scaled} to plot Series X and Y each scaled by their geometric mean
of the overlapping years (no scaling if no overlapping years), where X and
Y are A, B, C, or D.
\item \code{XY} to plot Series XY (with X rescaled in red), empty plot if longest
series is just X or Y - TODO - maybe. where X and
Y are A, B, C, or D.
}}

\item{x_lim}{x limits}

\item{y_lim}{y limits}

\item{shift}{if plotting two then shift Series X to left and Y to right by
\code{shift} amount}

\item{tck_length}{small tickmark lengths}

\item{ser_A_col}{colour for Series A}

\item{ser_B_col}{colour for Series B}

\item{ser_C_col}{colour for Series C}

\item{ser_D_col}{colour for Series D}

\item{legend_text}{text to override default if desired}

\item{x_lab}{x label to override default if desired}

\item{y_lab}{y label to override default if desired}

\item{gap_ci}{gap (units of 'height of letter O') between cirlce and
whiskers for confidence intervals}

\item{...}{further arguments to \code{gplots::plotCI}}
}
\value{

}
\description{
Single plotting function for Series A, B, C, D, or combination of two
}
\examples{
\dontrun{
# See vignettes.
}
}
\author{
Andrew Edwards
}
