% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-iphc.R
\name{plot_IPHC_ser_four_panels_ABCD}
\alias{plot_IPHC_ser_four_panels_ABCD}
\title{Wrapper to plot four versions of Series A-D plots in one figure,
with species name as main title.}
\usage{
plot_IPHC_ser_four_panels_ABCD(
  series_ABCD_full,
  sp = NULL,
  line_title = -2,
  y_lim = NULL,
  ...
)
}
\arguments{
\item{series_ABCD_full}{list containing tibbles for \code{ser_longest} etc.,
an output from \code{calc_iphc_full_res}}

\item{sp}{species common name}

\item{line_title}{for tweaking species title}

\item{y_lim}{y limits, if NULL (default) then automatically calculated}

\item{...}{further arguments to \code{plot_IPHC_ser_ABCD()}}
}
\value{
simple panel plot of four figures
}
\description{
Defaults to Series A, B, A and B scaled, and AB. Haven't generalised
\code{plot_IPHC_ser_ABCD()} to combinations with C and D yet (though does them
individually).
}
\author{
Andrew Edwards
}
