% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setData2020}
\alias{setData2020}
\title{Station data for 2020 IPHC survey}
\format{
A tibble:
\describe{
\item{year}{2020 for this data set}
\item{station}{station name}
\item{lat}{latitude of station}
\item{lon}{longitude of station}
\item{avgDepth}{average depth of set (in fathoms, haven't checked units for
earlier years)}
\item{effSkateIPHC}{effective skate number of the set, as calculated by IPHC}
\item{E_it20}{effective skate number of the set based on the first 20 hooks
only, calculated as the (number of hooks observed)/
(number of hooks retrieved) * effSkateIPHC}
\item{usable}{whether or not the set is usable, as determined by IPHC}
\item{standard}{whether or not the station is a standard station or in the
expanded grid}
\item{hooksRetr}{number of hooks retrieved for that set}
\item{hooksObs}{number of hooks observed for that set (already used to
calculate \code{E_it20}, but saving anyway)}
}
}
\source{
See \code{data-raw/iphc-2020-data.Rmd} for full details and how data were
extracted from IPHC website.
}
\usage{
setData2020
}
\description{
A dataset containing details of the stations for the 2020
IPHC survey; one set was deployed at each station except for station 2104,
but one vessel was \code{HAN} for which that was the only set in the survey, so it
has been removed.
}
\keyword{datasets}
