% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setDataExpansion}
\alias{setDataExpansion}
\title{Data indicating which stations are the 2018 (and maybe later) expansion stations}
\format{
A tibble:
\describe{
\item{station}{station name (character because early years used
alphanumeric and need to be able to combine years later)}
\item{lat}{latitude of station}
\item{lon}{longitude of station}
\item{standard}{whether or not (Y/N) the station is a standard one or an
expansions station}
\item{rca}{whether or not (Y/N) the station is within a Rockfish
Conservation Area}
\item{mpa}{whether or not (Y/N) the station is within a Marine Protected
Area}
\item{mpa_name}{name of the Marine Protected Area that the station is
within, if applicable (else is character <NA>)}
}
}
\source{
Originally from file IPHC_Stations_All.csv from Dana Haggarty (later
updated as IPHC_FISS_All_Stn_GFBio.csv to include the seven stations
originally missing but that were in GFbio), with
details and further tidying done here in \code{data-raw/IPHC-stations-expanded.R}.
}
\usage{
setDataExpansion
}
\description{
A dataset containing details of the stations for the 2018
IPHC survey; one set was deployed at each usual station, but an expanded grid (more
stations) was used this year, and the expansion stations need to be kept
track of when comparing spatially-averaged catch rates to previous
years. 2019 onwards may also use the expanded grids.
Effective skate number, avgDepth and usability will come from the GFbio query (unlike, say,
setData2013, for which the set details could not go into GFbio).
}
\keyword{datasets}
