% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sets_other_years}
\alias{sets_other_years}
\title{Details of sets for data extracted from GFBio}
\format{
A \code{tibble} with one row for each set, with columns:
\describe{
\item{year}{year}
\item{tripID}{trip ID number from GFBio}
\item{setID}{set ID number from GFBio}
\item{station}{IPHC station number}
\item{setInTrip}{set number during the trip}
\item{{long}{longitude}
\item{lat}{latitude}
\item{obsHooksPerSet}{observed number of hooks for that set}
\item{deplHooksPerSet}{deployed number of hooks for that set}
\item{skatesCount}{number of skates on the set}
\item{effSkateIPHC}{effective skate number from the IPHC}
\item{iphcUsabilityCode}{usability code from the IPHC}
\item{iphcUsabilityDesc}{usability description from the IPHC}
\item{usable}{whether or not (Y/N) the set is usable, as determined by the
IPHC}
\item{standard}{whether or not (Y/N) the set is in the standard area or the
expansion area (in 2018, and maybe later)}
}
}
\source{
Extracted from GFBio using \code{data-raw/sets-skates-hooks-yelloweye.R}.
}
\usage{
sets_other_years
}
\description{
Details of sets for data extracted from GFBio
}
\keyword{datasets}
