% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalise first letters of words (from ?toupper)}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{words to be capitalised}
}
\description{
Capitalise first letters of words (from ?toupper)
}
