% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-iphc-survey.R
\name{tidy_iphc_survey}
\alias{tidy_iphc_survey}
\title{Tidy IPHC survey data}
\usage{
tidy_iphc_survey(hook_level, skate_info, set_info)
}
\arguments{
\item{hook_level}{species-specific hook-level counts from \code{\link[=get_iphc_hooks]{get_iphc_hooks()}}.}

\item{skate_info}{species-independent skate-level information from
\code{\link[=get_iphc_skates_info]{get_iphc_skates_info()}}.}

\item{set_info}{species-indepentn set-level information from
\code{\link[=get_iphc_sets_info]{get_iphc_sets_info()}}.}
}
\value{
tibble with one row for each set in each year, with columns
year, station, lat, lon, E_it (effective skate number), N_it (number of
fish caught of the given species), C_it (catch rate of given species, as
numbers per effective skate), E_it20, N_it20 and C_it20 are the same but
considering the first 20 hooks of each skate only, usable (whether the set
should be used, based on IPHC codes; some cannot be used for geospatial
analysis but are included here). Returns single row with year 2003 and
all entries NA, so that functions that call this still work.
}
\description{
For a given species, calculate the catch rates for each skate based on all hooks
and based on the first 20 hooks only, for skates baited with chum (not the
other baits used in the 2012 bait experiment), for the years that we have
hook-by-hook data and all hooks were enumerated (2003-2012, 2014-2017 and
likely later).
}
\examples{
\dontrun{
yelloweye <- tidy_iphc_survey(
  get_iphc_hooks("yelloweye rockfish"),
  get_iphc_skates_info(),
  get_iphc_sets_info()
)
}
}
