% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{TODO_E_F_calc_iphc_full_res}
\alias{TODO_E_F_calc_iphc_full_res}
\title{Do all the calculations for the IPHC survey.}
\usage{
TODO_E_F_calc_iphc_full_res(set_counts)
}
\arguments{
\item{set_counts}{species-specific set-level counts from \code{\link[=tidy_iphc_survey]{tidy_iphc_survey()}}
or other.}
}
\value{
List containing

ser_longest: tibble for the longest time series that can be made for this
species, as output from \code{\link[=calc_iphc_ser_AB]{calc_iphc_ser_AB()}}; either Series A or AB.
Or Series B if a species is never caught when only the first 20 hooks
are enumerated (1997-2002 or 2013) but is caught when all hooks are
enumerated - then Series AB is just Series A but only includes years
that are in Series B, so we may as well use Series B (all hooks).
May need to think a little more. Just need results of test_BC to
be mentioned later.
full_coast: whether or not the longest time series can be considered
representative of the full coast (based on the paired t-tests).

ser_all: Series A, B, C and D, as output from \code{\link[=calc_iphc_ser_all]{calc_iphc_ser_all()}}.

test_AB: t-test results from \code{\link[=calc_iphc_ser_AB]{calc_iphc_ser_AB()}}

test_AD: t-test results from \code{\link[=compare_iphc_ser_A_D]{compare_iphc_ser_A_D()}}

test_BC: t-test results from \code{\link[=compare_iphc_ser_B_C]{compare_iphc_ser_B_C()}}

If no observations at all for the species then return NA (says NULL, no?).
}
\description{
From species-specific set-level counts, derive all the required Series and
test their equivalence.
}
