% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser-area.R
\name{TODO_E_F_iphc_get_calc_plot}
\alias{TODO_E_F_iphc_get_calc_plot}
\title{Get data, do calculations and plot longest series for the IPHC survey}
\usage{
TODO_E_F_iphc_get_calc_plot(sp)
}
\arguments{
\item{sp}{Species names (as used in gfdata and gfplot).}
}
\value{
For the given species, list containing

iphc_set_counts_sp: list returned from \code{\link[=calc_iphc_full_res]{calc_iphc_full_res()}}

iphc_set_counts_sp_format: just the longest series, returned from
\code{\link[=format_iphc_longest]{format_iphc_longest()}} with calculations and formatting to match that
of other surveys

g_iphc_index: plot of just the iphc data (useful for testing all species)
providing there are some data
}
\description{
Get data, do calculations and plot longest series for the IPHC survey for
a given species. Will take a while since queries GFbio (and need to be on DFO
network). WON'T CURRENTLY WORK AS plot_iphc_index() won't currently work.
}
