% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yelloweye_rockfish}
\alias{yelloweye_rockfish}
\title{Available counts of Yelloweye Rockfish, for each set for all years}
\format{
A \code{list} containing a single \code{tibble} called \code{set_counts}, with
columns as described in \code{?get_all_iphc_set_counts}.
}
\source{
Extracted from GFBio using \code{data-raw/sets-skates-hooks-yelloweye.R}.
}
\usage{
yelloweye_rockfish
}
\description{
Extracts and calculates all avaiable counts, at the set level, of Yelloweye
Rockfish, as an example dataset (see vignettes).
}
\keyword{datasets}
