% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{iphc_get_calc_plot_full}
\alias{iphc_get_calc_plot_full}
\title{Get data, do calculations and plot Series A, B and longest series possible
(usually Series AB) for the IPHC survey, and maybe save results}
\usage{
iphc_get_calc_plot_full(
  sp,
  cached = TRUE,
  verbose = FALSE,
  print_sp_name = TRUE,
  path_data = ".",
  path_results = NULL
)
}
\arguments{
\item{sp}{Species names (as used in gfdata and gfplot). Or something like
"skates combined" -- see vignette.}

\item{cached}{if TRUE then used cached data (sp-name.rds)}

\item{verbose}{if TRUE then print out some of the data (useful in vignette loops)}

\item{print_sp_name}{if TRUE then print out species name (useful in vignette
loops)}

\item{path_data}{path to save or load the cached data}

\item{path_results}{path to save or load the cached data}
}
\value{
Saves results in \code{path_results/species-name.RDS} if \code{path_results}
is not NULL. For the given species, return list containing

sp_set_counts: list with one element (for consistency), a tibble
\code{set_counts} (the .RDS file saved when doing \code{cache_pbs_data_iphc(sp)}).

ser_ABCD_full: list of output from \code{calc_iphc_full_res()} from doing all
calculations.
}
\description{
Get data, do calculations and plot longest series for the IPHC survey for
a given species. For species that have several years of zero catches,
Series B may be better to use (but shorter) than Series AB because all
hooks were counted. For example Walleye Pollock in 2016 -- see example.
}
\details{
Will take a while since queries GFbio (and need to be on DFO
network). Basically a wrapper for the calculations in the vignette
\code{data_for_one_species}. Based on \code{iphc_get_calc_plot_area()}. See
\code{gfsynopsis::iphc_get_calc_plot()} for function for gfsynopsis
reports. Calling this one \code{iphc_get_calc_plot_full()} to distinguish it.
}
\author{
Andrew Edwards
}
