% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{plot_iphc_index}
\alias{plot_iphc_index}
\title{Plot just the IPHC survey index, though works for all surveys}
\usage{
plot_iphc_index(iphc_set_counts_sp_format)
}
\arguments{
\item{iphc_set_counts_sp_format}{Set counts for a given species formatted
in the same way as the other survey series, using \code{\link[=format_iphc_longest]{format_iphc_longest()}}}
}
\value{
ggplot object of the plot
}
\description{
Plot just the IPHC survey index (mainly for testing).
}
