% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphc-maps.R
\name{plot_iphc_map}
\alias{plot_iphc_map}
\title{Plot the locations of the IPHC stations for a given year, and whether or not
a given species was caught in that year.}
\usage{
plot_iphc_map(
  set_counts_of_sp,
  sp_short_name = NULL,
  years,
  main_title = NULL,
  mar_val = c(1.8, 2, 1, 0.5),
  mgp_val = c(1.6, 0.5, 0),
  lat_cut_off = 50.6,
  lon_cut_off_1 = -130,
  lon_cut_off_2 = -128.25,
  pch_pos_count = 19,
  pch_zero_count = 1,
  cex_val = 1,
  ...
)
}
\arguments{
\item{set_counts_of_sp}{input tibble of the species of interest, likely generated from
\code{cache_pbs_data_iphc} (which seems to save a list with the first element
the desired tibble), with column names \code{year}, \code{station}, \code{lat}, \code{lon},
\code{E_it}, \code{N_it}, \code{C_it}, \code{E_it20}, \code{N_it20}, \code{C_it20}, \code{usable}}

\item{sp_short_name}{short name species of interest to have in legend,
if NULL then plots all stations labelled as
usable and unusable, but with no catch rates. Still needs
\code{set_counts_of_sp} input tibble (can be any species, as they all have all
stations documented).}

\item{years}{years of interest ****currently just for one year, multiple will
get tricky}

\item{main_title}{title for map, if NULL then is "All <year> stations"}

\item{mar_val}{mar values to reduce whitespace around maps}

\item{mgp_val}{mgp values to reduce whitespace around maps}

\item{lat_cut_off}{latitudinal cut off near top of Vancouver Island, below
which stations are excluded when constructing Series A and B; this is just
below all stations from 1995-1998, so that Series A and B include all
stations for 1995-1998}

\item{lon_cut_off_1}{left-hand end of line (longitude) to show cut off}

\item{lon_cut_off_2}{right-hand end of line (longitude) to show cut off}

\item{pch_pos_count}{pch for positive counts}

\item{pch_zero_count}{pch for positive counts}

\item{cex_val}{cex size of plotted circles (or whatever is chosen using
above \code{pch} options}
}
\value{
A map of the IPHC survey stations for that year and species, with a
legend describing the points.
}
\description{
If possible, also show whether the species was caught in the first 20 hooks
and only caught after the first 20 hooks. MAY ADD other options. Uses
\code{PBSmapping} style of plot, as used in the Redbanded and Yelloweye Rockfish
stock assessments.
}
\author{
Andrew Edwards
}
