% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-iphc.R
\name{plot_IPHC_ser_four_panels}
\alias{plot_IPHC_ser_four_panels}
\title{Wrapper to plot all four versions of Series E and F plots in one figure,
with species name as main title}
\usage{
plot_IPHC_ser_four_panels(
  ser_E_and_F,
  series_longest,
  sp = NULL,
  line_title = -2,
  ...
)
}
\arguments{
\item{ser_E_and_F}{list containing tibbles for \code{ser_E} and \code{ser_F} of class
\code{IPHC_ser_E_and_F}, an output from \code{calc_iphc_ser_E_and_F}}

\item{series_longest}{list of outputs from \code{calc_iphc_ser_EF(ser_E_and_F)},
including tibble \code{ser_longest} plus t-test results and geometric means of
each series.}

\item{sp}{species common name}

\item{line_title}{line number to put the title, tweak if necessary}

\item{...}{further arguments to \code{plot_IPHC_ser_E_and_F()}}
}
\value{
simple panel plot of four figures
}
\description{
Wrapper to plot all four versions of Series E and F plots in one figure,
with species name as main title
}
\examples{
\dontrun{
# see vignette
}
}
\author{
Andrew Edwards
}
